Move `maskFragments` and `unsafe_readResult` into `gql.tada/testing`
