+++ src/api.ts
@@ -14,14 +14,13 @@ import type {
   FragmentShape,
   getFragmentsOfDocuments,
   decorateFragmentDef,
-  omitFragmentRefsRec,
   makeFragmentRef,
 } from './namespace';
 
 import type { getDocumentType } from './selection';
 import type { parseDocument, DocumentNodeLike } from './parser';
 import type { getVariablesType, getScalarType } from './variables';
-import type { obj, matchOr, writable, DocumentDecoration } from './utils';
+import type { matchOr, writable, DocumentDecoration } from './utils';
 import { concatLocSources } from './utils';
 
 /** Abstract configuration type input for your schema and scalars.
@@ -513,20 +512,6 @@ type resultOfT<Document extends FragmentShape, T = unknown> =
       : Result
     : never;
 
-type resultOfFragmentsRec<
-  Fragments extends readonly any[],
-  Result = {},
-> = Fragments extends readonly [infer Fragment, ...infer Rest]
-  ? resultOfFragmentsRec<Rest, ResultOf<Fragment> & Result>
-  : Result;
-
-type fragmentRefsOfFragmentsRec<
-  Fragments extends readonly any[],
-  FragmentRefs = {},
-> = Fragments extends readonly [infer Fragment, ...infer Rest]
-  ? fragmentRefsOfFragmentsRec<Rest, makeFragmentRef<Fragment> & FragmentRefs>
-  : obj<FragmentRefs>;
-
 /** Unmasks a fragment mask for a given fragment document and data.
  *
  * @param _document - A GraphQL document of a fragment, created using {@link graphql}.
@@ -652,121 +637,10 @@ function readFragment(...args: [unknown] | [unknown, unknown]) {
   return args.length === 2 ? args[1] : args[0];
 }
 
-/** For testing, masks fragment data for given data and fragments.
- *
- * @param _fragments - A list of GraphQL documents of fragments, created using {@link graphql}.
- * @param data - The combined result data of the fragments, which can be wrapped in arrays.
- * @returns The masked data of the fragments.
- *
- * @remarks
- * When creating test data, you may define data for fragments that’s unmasked, making it
- * unusable in parent fragments or queries that require masked data.
- *
- * This means that you may have to use {@link maskFragments} to mask your data first
- * for TypeScript to not report an error.
- *
- * @example
- * ```
- * import { FragmentOf, ResultOf, graphql, maskFragments } from 'gql.tada';
- *
- * const bookFragment = graphql(`
- *   fragment BookComponent on Book {
- *     id
- *     title
- *   }
- * `);
- *
- * const data = maskFragments([bookFragment], { id: 'id', title: 'book' });
- * ```
- *
- * @see {@link readFragment} for how to read from fragment masks (i.e. the reverse)
- */
-function maskFragments<const Fragments extends readonly FragmentShape[]>(
-  _fragments: Fragments,
-  fragment: resultOfFragmentsRec<Fragments>
-): fragmentRefsOfFragmentsRec<Fragments>;
-function maskFragments<const Fragments extends readonly FragmentShape[]>(
-  _fragments: Fragments,
-  fragment: resultOfFragmentsRec<Fragments> | null
-): fragmentRefsOfFragmentsRec<Fragments> | null;
-function maskFragments<const Fragments extends readonly FragmentShape[]>(
-  _fragments: Fragments,
-  fragment: resultOfFragmentsRec<Fragments> | undefined
-): fragmentRefsOfFragmentsRec<Fragments> | undefined;
-function maskFragments<const Fragments extends readonly FragmentShape[]>(
-  _fragments: Fragments,
-  fragment: resultOfFragmentsRec<Fragments> | null | undefined
-): fragmentRefsOfFragmentsRec<Fragments> | null | undefined;
-function maskFragments<const Fragments extends readonly FragmentShape[]>(
-  _fragments: Fragments,
-  fragment: readonly resultOfFragmentsRec<Fragments>[]
-): readonly fragmentRefsOfFragmentsRec<Fragments>[];
-function maskFragments<const Fragments extends readonly FragmentShape[]>(
-  _fragments: Fragments,
-  fragment: readonly (resultOfFragmentsRec<Fragments> | null)[]
-): readonly (fragmentRefsOfFragmentsRec<Fragments> | null)[];
-function maskFragments<const Fragments extends readonly FragmentShape[]>(
-  _fragments: Fragments,
-  fragment: readonly (resultOfFragmentsRec<Fragments> | undefined)[]
-): readonly (fragmentRefsOfFragmentsRec<Fragments> | undefined)[];
-function maskFragments<const Fragments extends readonly FragmentShape[]>(
-  _fragments: Fragments,
-  fragment: readonly (resultOfFragmentsRec<Fragments> | null | undefined)[]
-): readonly (fragmentRefsOfFragmentsRec<Fragments> | null | undefined)[];
-function maskFragments(_fragments: unknown, data: unknown) {
-  return data;
-}
-
-/** For testing, converts document data without fragment refs to their result type.
- *
- * @param _document - A GraphQL document, created using {@link graphql}.
- * @param data - The result data of the GraphQL document with optional fragment refs.
- * @returns The masked result data of the document.
- *
- * @remarks
- * When creating test data, you may define data for documents that’s unmasked, but
- * need to cast the data to match the result type of your document.
- *
- * This means that you may have to use {@link unsafe_readResult} to cast
- * them to the result type, instead of doing `as any as ResultOf<typeof document>`.
- *
- * This function is inherently unsafe, since it doesn't check that your document
- * actually contains the masked fragment data!
- *
- * @example
- * ```
- * import { FragmentOf, ResultOf, graphql, unsafe_readResult } from 'gql.tada';
- *
- * const bookFragment = graphql(`
- *   fragment BookComponent on Book {
- *     id
- *     title
- *   }
- * `);
- *
- * const query = graphql(`
- *   query {
- *     book {
- *       ...BookComponent
- *     }
- *   }
- * `, [bookFragment]);
- *
- * const data = unsafe_readResult(query, { book: { id: 'id', title: 'book' } });
- * ```
- *
- * @see {@link readFragment} for how to read from fragment masks (i.e. the reverse)
- */
-function unsafe_readResult<
-  const Document extends DocumentDecoration<any, any>,
-  const Data extends omitFragmentRefsRec<ResultOf<Document>>,
->(_document: Document, data: Data): ResultOf<Document> {
-  return data as any;
-}
-
 const graphql: initGraphQLTada<setupSchema> = initGraphQLTada();
 
-export { parse, graphql, readFragment, maskFragments, unsafe_readResult };
+export { parse, graphql, readFragment };
+export { maskFragments, unsafe_readResult } from './testing';
 
 export type {
   setupCache,
